+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog]
   - Tasks that have passed also show in the 'view claims' section
   - Update the claims index view to display status of identity and qualification
     tasks if they have been manually passed/failed
+  - Display stats for outcomes of School Workforce Census checks
 
 - Update documentation